+++ man-roxygen/param_r_max.R
@@ -1,0 +1,2 @@
+#' @param r_max (`numeric(1)`)\cr
+#' Upper bound of budget parameter.